+++ src/intelligence_layer/connectors/__init__.py
@@ -49,6 +49,9 @@ from .document_index.document_index import DocumentSearchResult as DocumentSearc
 from .document_index.document_index import (
     ExternalServiceUnavailable as ExternalServiceUnavailable,
 )
+from .document_index.document_index import FilterField as FilterField
+from .document_index.document_index import FilterOps as FilterOps
+from .document_index.document_index import Filters as Filters
 from .document_index.document_index import IndexConfiguration as IndexConfiguration
 from .document_index.document_index import IndexPath as IndexPath
 from .document_index.document_index import InternalError as InternalError