+++ tests/connectors/document_index/test_document_index.py
@@ -1,15 +1,23 @@
+from datetime import datetime, timezone
 from http import HTTPStatus
 
 import pytest
+from pydantic import ValidationError
 from pytest import fixture, raises
 
+from intelligence_layer.connectors.base.json_serializable import JsonSerializable
 from intelligence_layer.connectors.document_index.document_index import (
     CollectionPath,
     DocumentContents,
     DocumentFilterQueryParams,
     DocumentIndexClient,
     DocumentPath,
+    FilterField,
+    FilterOps,
+    Filters,
+    IndexConfiguration,
     IndexPath,
+    InvalidInput,
     ResourceNotFound,
     SearchQuery,
 )
@@ -18,6 +26,32 @@ from intelligence_layer.connectors.document_index.document_index import (
 @fixture
 def aleph_alpha_namespace() -> str:
     return "aleph-alpha"
+
+
+@fixture
+def filter_index_config() -> dict[str, dict[str, str]]:
+    return {
+        "test-string-filter": {
+            "field-name": "string-field",
+            "field-type": "string",
+        },
+        "test-integer-filter": {
+            "field-name": "integer-field",
+            "field-type": "integer",
+        },
+        "test-float-filter": {
+            "field-name": "float-field",
+            "field-type": "float",
+        },
+        "test-boolean-filter": {
+            "field-name": "boolean-field",
+            "field-type": "boolean",
+        },
+        "test-date-filter": {
+            "field-name": "date-field",
+            "field-type": "date_time",
+        },
+    }
 
 
 @fixture
@@ -45,6 +79,49 @@ Under Candler's leadership, Coca-Cola transformed from a pharmacist's concoction
 
 Pemberton's life story is a testament to the spirit of innovation and resilience. His creation, borne out of personal struggles and the context of his times, went on to transcend its origins and become a symbol recognized across the globe. Today, when we think of Coca-Cola, we are reminded of Pemberton's journey from a small-town pharmacist to the creator of one of the world's most enduring and beloved brands."""
     return DocumentContents(contents=[text], metadata={"Some": "Metadata"})
+
+
+@fixture
+def document_contents_with_metadata() -> list[DocumentContents]:
+    text_1 = """John Stith Pemberton, the inventor of the world-renowned beverage Coca-Cola, was a figure whose life was marked by creativity, entrepreneurial spirit, and the turbulent backdrop of 19th-century America. Born on January 8, 1831, in Knoxville, Georgia, Pemberton grew up in an era of profound transformation and change."""
+    text_2 = """Pemberton began his professional journey by studying medicine and pharmacy. After earning a degree in pharmacy, he started his career as a druggist in Columbus, Georgia. He was known for his keen interest in creating medicinal concoctions and was well-respected in his community. His early creations included various medicines and tonics, which were typical of the times when pharmacists often concocted their own remedies."""
+    text_3 = """Pemberton's life took a significant turn during the American Civil War. He served as a lieutenant colonel in the Confederate Army, and it was during this period that he sustained a wound that led him to become dependent on morphine. This personal struggle with addiction likely influenced his later work in seeking out alternatives and remedies for pain relief."""
+
+    metadata_1: JsonSerializable = {
+        "string-field": "example_string_1",
+        "integer-field": 123,
+        "float-field": 123.45,
+        "boolean-field": True,
+        "date-field": datetime(2022, 1, 1, tzinfo=timezone.utc)
+        .isoformat(timespec="seconds")
+        .replace("+00:00", "Z"),
+    }
+
+    metadata_2: JsonSerializable = {
+        "string-field": "example_string_2",
+        "integer-field": 456,
+        "float-field": 678.90,
+        "boolean-field": False,
+        "date-field": datetime(2023, 1, 1, tzinfo=timezone.utc)
+        .isoformat(timespec="seconds")
+        .replace("+00:00", "Z"),
+    }
+
+    metadata_3: JsonSerializable = {
+        "string-field": "example_string_3",
+        "integer-field": 789,
+        "float-field": 101112.13,
+        "boolean-field": True,
+        "date-field": datetime(2024, 1, 1, tzinfo=timezone.utc)
+        .isoformat(timespec="seconds")
+        .replace("+00:00", "Z"),
+    }
+
+    return [
+        DocumentContents(contents=[text_1], metadata=metadata_1),
+        DocumentContents(contents=[text_2], metadata=metadata_2),
+        DocumentContents(contents=[text_3], metadata=metadata_3),
+    ]
 
 
 @pytest.mark.internal
@@ -196,7 +273,7 @@ def test_document_list_all_documents(
 ) -> None:
     filter_result = document_index.documents(collection_path)
 
-    assert len(filter_result) == 3
+    assert len(filter_result) == 6
 
 
 def test_document_list_max_n_documents(
@@ -233,6 +310,17 @@ def test_document_path_is_immutable() -> None:
     assert dictionary[path] == 1
 
 
+def test_index_configuration_rejects_invalid_chunk_overlap() -> None:
+    try:
+        IndexConfiguration(
+            chunk_size=128, chunk_overlap=128, embedding_type="asymmetric"
+        )
+    except ValidationError as e:
+        assert "chunk_overlap must be less than chunk_size" in str(e)
+    else:
+        raise AssertionError("ValidationError was not raised")
+
+
 def test_document_indexes_are_returned(
     document_index: DocumentIndexClient, collection_path: CollectionPath
 ) -> None:
@@ -243,4 +331,382 @@ def test_document_indexes_are_returned(
     )
 
     assert index_configuration.embedding_type == "asymmetric"
+    assert index_configuration.chunk_overlap == 0
     assert index_configuration.chunk_size == 512
+
+
+def test_create_filter_indexes_in_namespace(
+    document_index: DocumentIndexClient,
+    aleph_alpha_namespace: str,
+    filter_index_config: dict[str, dict[str, str]],
+) -> None:
+    for index_name, index_config in filter_index_config.items():
+        document_index.create_filter_index_in_namespace(
+            namespace=aleph_alpha_namespace,
+            filter_index_name=index_name,
+            field_name=index_config["field-name"],
+            field_type=index_config["field-type"],  # type:ignore[arg-type]
+        )
+
+    assert all(
+        filter_index
+        in document_index.list_filter_indexes_in_namespace(aleph_alpha_namespace)
+        for filter_index in filter_index_config
+    )
+
+
+def test_create_filter_index_invalid_name(
+    document_index: DocumentIndexClient, aleph_alpha_namespace: str
+) -> None:
+    with pytest.raises(ValueError) as context:
+        document_index.create_filter_index_in_namespace(
+            aleph_alpha_namespace, "invalid index!", "field_name", "string"
+        )
+        assert (
+            str(context.value)
+            == "Filter index name can only contain alphanumeric characters (a-z, A-Z, -, . and 0-9)."
+        )
+
+
+def test_create_filter_index_name_too_long(
+    document_index: DocumentIndexClient, aleph_alpha_namespace: str
+) -> None:
+    with pytest.raises(ValueError) as context:
+        document_index.create_filter_index_in_namespace(
+            aleph_alpha_namespace, "a" * 51, "field_name", "string"
+        )
+    assert (
+        str(context.value) == "Filter index name cannot be longer than 50 characters."
+    )
+
+
+def test_assign_filter_indexes_to_collection(
+    document_index: DocumentIndexClient,
+    collection_path: CollectionPath,
+    filter_index_config: dict[str, dict[str, str]],
+) -> None:
+    for index_name in filter_index_config:
+        document_index.assign_filter_index_to_search_index(
+            collection_path=collection_path,
+            filter_index_name=index_name,
+            index_name="asymmetric",
+        )
+
+    assert all(
+        filter_index
+        in document_index.list_assigned_filter_index_names(
+            collection_path, "asymmetric"
+        )
+        for filter_index in filter_index_config
+    )
+
+
+def test_document_index_adds_documents_with_metadata(
+    document_index: DocumentIndexClient,
+    collection_path: CollectionPath,
+    document_contents_with_metadata: list[DocumentContents],
+) -> None:
+    for i, doc_content in enumerate(document_contents_with_metadata):
+        document_path = DocumentPath(
+            collection_path=collection_path,
+            document_name=f"document-metadata-{i}",
+        )
+        document_index.add_document(document_path, doc_content)
+
+        assert any(
+            d.document_path == document_path
+            for d in document_index.documents(collection_path)
+        )
+        assert doc_content == document_index.document(document_path)
+
+
+def test_search_with_string_filter(
+    document_index: DocumentIndexClient,
+    collection_path: CollectionPath,
+) -> None:
+    search_query = SearchQuery(
+        query="Coca-Cola",
+        max_results=10,
+        min_score=0.5,
+        filters=[
+            Filters(
+                filter_type="with",
+                fields=[
+                    FilterField(
+                        field_name="string-field",
+                        field_value="example_string_1",
+                        criteria=FilterOps.EQUAL_TO,
+                    )
+                ],
+            )
+        ],
+    )
+    results = document_index.search(collection_path, "asymmetric", search_query)
+    assert results[0].document_path.document_name == "document-metadata-0"
+
+
+def test_search_with_integer_filter(
+    document_index: DocumentIndexClient,
+    collection_path: CollectionPath,
+) -> None:
+    search_query = SearchQuery(
+        query="Coca-Cola",
+        max_results=10,
+        min_score=0.5,
+        filters=[
+            Filters(
+                filter_type="with",
+                fields=[
+                    FilterField(
+                        field_name="integer-field",
+                        field_value=123,
+                        criteria=FilterOps.EQUAL_TO,
+                    )
+                ],
+            )
+        ],
+    )
+    results = document_index.search(collection_path, "asymmetric", search_query)
+    assert len(results) == 1
+    assert results[0].document_path.document_name == "document-metadata-0"
+
+
+def test_search_with_float_filter(
+    document_index: DocumentIndexClient,
+    collection_path: CollectionPath,
+) -> None:
+    search_query = SearchQuery(
+        query="Coca-Cola",
+        max_results=10,
+        min_score=0.1,
+        filters=[
+            Filters(
+                filter_type="with",
+                fields=[
+                    FilterField(
+                        field_name="float-field",
+                        field_value=123.45,
+                        criteria=FilterOps.GREATER_THAN,
+                    )
+                ],
+            )
+        ],
+    )
+    results = document_index.search(collection_path, "asymmetric", search_query)
+    assert len(results) == 2
+    assert results[0].document_path.document_name == "document-metadata-1"
+    assert results[1].document_path.document_name == "document-metadata-2"
+
+
+def test_search_with_boolean_filter(
+    document_index: DocumentIndexClient,
+    collection_path: CollectionPath,
+) -> None:
+    search_query = SearchQuery(
+        query="Coca-Cola",
+        max_results=10,
+        min_score=0.5,
+        filters=[
+            Filters(
+                filter_type="with",
+                fields=[
+                    FilterField(
+                        field_name="boolean-field",
+                        field_value=True,
+                        criteria=FilterOps.EQUAL_TO,
+                    )
+                ],
+            )
+        ],
+    )
+    results = document_index.search(collection_path, "asymmetric", search_query)
+    assert len(results) == 1
+    assert results[0].document_path.document_name == "document-metadata-0"
+
+
+def test_search_with_datetime_filter(
+    document_index: DocumentIndexClient,
+    collection_path: CollectionPath,
+) -> None:
+    search_query = SearchQuery(
+        query="Coca-Cola",
+        max_results=10,
+        min_score=0.1,
+        filters=[
+            Filters(
+                filter_type="with",
+                fields=[
+                    FilterField(
+                        field_name="date-field",
+                        field_value=datetime(2023, 1, 1, tzinfo=timezone.utc),
+                        criteria=FilterOps.BEFORE,
+                    )
+                ],
+            )
+        ],
+    )
+    results = document_index.search(collection_path, "asymmetric", search_query)
+    assert len(results) == 1
+    assert results[0].document_path.document_name == "document-metadata-0"
+
+
+def test_search_with_invalid_datetime_filter(
+    document_index: DocumentIndexClient,
+    collection_path: CollectionPath,
+) -> None:
+    search_query = SearchQuery(
+        query="Coca-Cola",
+        max_results=10,
+        min_score=0.1,
+        filters=[
+            Filters(
+                filter_type="with",
+                fields=[
+                    FilterField(
+                        field_name="date-field",
+                        field_value="2023-01-01T12:00:00",
+                        criteria=FilterOps.BEFORE,
+                    )
+                ],
+            )
+        ],
+    )
+    with raises(InvalidInput):
+        document_index.search(collection_path, "asymmetric", search_query)
+
+
+def test_search_with_multiple_filters(
+    document_index: DocumentIndexClient,
+    collection_path: CollectionPath,
+) -> None:
+    search_query = SearchQuery(
+        query="Coca-Cola",
+        max_results=10,
+        min_score=0.1,
+        filters=[
+            Filters(
+                filter_type="with",
+                fields=[
+                    FilterField(
+                        field_name="integer-field",
+                        field_value=123,
+                        criteria=FilterOps.EQUAL_TO,
+                    ),
+                    FilterField(
+                        field_name="boolean-field",
+                        field_value=True,
+                        criteria=FilterOps.EQUAL_TO,
+                    ),
+                ],
+            )
+        ],
+    )
+    results = document_index.search(collection_path, "asymmetric", search_query)
+    assert len(results) == 1
+    assert results[0].document_path.document_name == "document-metadata-0"
+
+
+def test_search_with_filter_type_without(
+    document_index: DocumentIndexClient,
+    collection_path: CollectionPath,
+) -> None:
+    search_query = SearchQuery(
+        query="Coca-Cola",
+        max_results=10,
+        min_score=0.1,
+        filters=[
+            Filters(
+                filter_type="without",
+                fields=[
+                    FilterField(
+                        field_name="integer-field",
+                        field_value=456,
+                        criteria=FilterOps.EQUAL_TO,
+                    )
+                ],
+            )
+        ],
+    )
+    results = document_index.search(collection_path, "asymmetric", search_query)
+    assert len(results) == 7
+
+
+def test_search_with_filter_type_without_and_with(
+    document_index: DocumentIndexClient,
+    collection_path: CollectionPath,
+) -> None:
+    search_query = SearchQuery(
+        query="Coca-Cola",
+        max_results=10,
+        min_score=0.1,
+        filters=[
+            Filters(
+                filter_type="without",
+                fields=[
+                    FilterField(
+                        field_name="integer-field",
+                        field_value=456,
+                        criteria=FilterOps.EQUAL_TO,
+                    )
+                ],
+            ),
+            Filters(
+                filter_type="with",
+                fields=[
+                    FilterField(
+                        field_name="boolean-field",
+                        field_value=True,
+                        criteria=FilterOps.EQUAL_TO,
+                    )
+                ],
+            ),
+        ],
+    )
+    results = document_index.search(collection_path, "asymmetric", search_query)
+    assert len(results) == 2
+    assert results[0].document_path.document_name == "document-metadata-0"
+    assert results[1].document_path.document_name == "document-metadata-2"
+
+
+def test_search_with_filter_type_with_one_of(
+    document_index: DocumentIndexClient,
+    collection_path: CollectionPath,
+) -> None:
+    search_query = SearchQuery(
+        query="Coca-Cola",
+        max_results=10,
+        min_score=0.1,
+        filters=[
+            Filters(
+                filter_type="with_one_of",
+                fields=[
+                    FilterField(
+                        field_name="integer-field",
+                        field_value=456,
+                        criteria=FilterOps.EQUAL_TO,
+                    )
+                ],
+            ),
+            Filters(
+                filter_type="with_one_of",
+                fields=[
+                    FilterField(
+                        field_name="integer-field",
+                        field_value=789,
+                        criteria=FilterOps.EQUAL_TO,
+                    )
+                ],
+            ),
+        ],
+    )
+    results = document_index.search(collection_path, "asymmetric", search_query)
+    assert len(results) == 2
+    assert results[0].document_path.document_name == "document-metadata-1"
+    assert results[1].document_path.document_name == "document-metadata-2"
+
+
+def test_document_indexes_zero_progress_is_returned(
+    document_index: DocumentIndexClient, collection_path: CollectionPath
+) -> None:
+    progress = document_index.progress(collection_path)
+    assert progress == 0